+++ Documentation/Presentation Host Examples and Tests/3. Externally Dimissed.swift
@@ -1,0 +1,27 @@
+import SwiftUI
+import VCore
+
+struct ContentView: View {
+    @State private var isPresented: Bool = false
+
+    var body: some View {
+        Button(
+            "Present",
+            action: {
+                isPresented = true
+
+                Task(operation: {
+                    try? await Task.sleep(seconds: 1)
+                    isPresented = false
+                })
+            }
+        )
+        .someModal(
+            id: "some_modal",
+            isPresented: $isPresented,
+            content: { Color.accentColor }
+        )
+    }
+}
+
+// ...